moving definition arguments to a new module
