+++ src/task_definition/definition_arguments_test.rs
@@ -1,0 +1,4 @@
+#[test]
+fn test_() {
+    assert_eq!(1, 2);
+}
