+++ src/task_definition/mod.rs
@@ -3,6 +3,7 @@ mod dummy_task;
 mod python_task;
 mod task_def;
 mod definition_factory;
+mod definition_arguments;
 
 pub use bash_task::BashTaskDefinition;
 pub use dummy_task::DummyTaskDefinition;
@@ -10,6 +11,9 @@ pub use python_task::PythonTaskDefinition;
 pub use task_def::{
     generate_task_definition_id, TaskDefinition
 };
+pub use definition_arguments::{
+    DefinitionArguments, DefinitionArgumentType, DefinitionArgumentElement
+};
 pub use definition_factory::{
-    string_to_definition_type, TaskDefinitionType, DefinitionArguments, DefinitionArgumentType, DefinitionArgumentElement
+    string_to_definition_type, TaskDefinitionType
 };
